Separate template mobile and panel desktop with mixin.
